+++ server/src/lib/mongodb.js
@@ -7,7 +7,7 @@ Mongo = {}
 
 Mongo.query=async function(connection_name,query,callback){
   // does the pool exist already, if not let's add it
-  logger.debug("["+connection_name+"] running query : " + query)
+  // logger.debug("["+connection_name+"] running query : " + query)
   var config=undefined
   var queryarr = query.split("~")
   if(queryarr.length!=3){
@@ -18,7 +18,7 @@ Mongo.query=async function(connection_name,query,callback){
   try{
     var queryprep = JSON.parse(queryarr[2])
   }catch(err){
-    logger.error("["+connection_name+"] query must valid json. Use double quotes, not single quotes.  " + queryarr[2])
+    logger.error("["+connection_name+"] query must be valid json. Use double quotes, not single quotes.  " + queryarr[2])
     callback(null,null)
     return;
   }
@@ -31,7 +31,7 @@ Mongo.query=async function(connection_name,query,callback){
   }
 
   var uri = `mongodb://${encodeURI(config.user)}:${encodeURI(config.password)}@${config.host}:${config.port}`
-  logger.silly("["+connection_name+"] uri : " + uri)
+  // logger.silly("["+connection_name+"] uri : " + uri)
   try{
     Client.connect(uri,function(err,db){
       if(err){
@@ -41,7 +41,7 @@ Mongo.query=async function(connection_name,query,callback){
         var dbo=db.db(queryarr[0])
 
         dbo.collection(queryarr[1]).find(queryprep).toArray(function(err,result){
-          logger.silly("["+connection_name+"] query result : " + JSON.stringify(result))
+          // logger.silly("["+connection_name+"] query result : " + JSON.stringify(result))
           callback(null,result)
           db.close()
         })